Working code

+++ include/geometry/helpers.hpp
@@ -37,7 +37,7 @@ int orientation(Point p, Point q, Point r)
 {
 	// See https://www.geeksforgeeks.org/orientation-3-ordered-points/
 	// for details of below formula.
-	double val = (q.y - p.y) * (r.x - q.x) -
+	float val = (q.y - p.y) * (r.x - q.x) -
 		(q.x - p.x) * (r.y - q.y);
 
 	if (std::abs(val) < eps) return 0;  // collinear
@@ -48,9 +48,9 @@ int orientation(Point p, Point q, Point r)
 /// <summary>
 /// Function to find if 2 line segments intersect.
 /// </summary>
-/// <param name="p1">1st endpoint of the 1st line segment.</param>
+/// <param name="p_1">1st endpoint of the 1st line segment.</param>
 /// <param name="q1">2nd endpoint of the 1st line segment.</param>
-/// <param name="p2">1st endpoint of the 2nd line segment.</param>
+/// <param name="p_2">1st endpoint of the 2nd line segment.</param>
 /// <param name="q2">2nd endpoint of the 2nd line segment.</param>
 /// <returns>True, if the line segments intersect; False if they don't.</returns>
 bool doIntersect(Point p1, Point q1, Point p2, Point q2)
@@ -67,16 +67,16 @@ bool doIntersect(Point p1, Point q1, Point p2, Point q2)
 		return true;
 
 	// Special Cases
-	// p1, q1 and p2 are collinear and p2 lies on segment p1q1
+	// p_1, q1 and p_2 are collinear and p_2 lies on segment p1q1
 	if (o1 == 0 && onSegment(p1, p2, q1)) return true;
 
-	// p1, q1 and q2 are collinear and q2 lies on segment p1q1
+	// p_1, q1 and q2 are collinear and q2 lies on segment p1q1
 	if (o2 == 0 && onSegment(p1, q2, q1)) return true;
 
-	// p2, q2 and p1 are collinear and p1 lies on segment p2q2
+	// p_2, q2 and p_1 are collinear and p_1 lies on segment p2q2
 	if (o3 == 0 && onSegment(p2, p1, q2)) return true;
 
-	// p2, q2 and q1 are collinear and q1 lies on segment p2q2
+	// p_2, q2 and q1 are collinear and q1 lies on segment p2q2
 	if (o4 == 0 && onSegment(p2, q1, q2)) return true;
 
 	return false; // Doesn't fall in any of the above cases
@@ -93,26 +93,26 @@ bool doIntersect(Point p1, Point q1, Point p2, Point q2)
 Point intersection(Point A, Point B, Point C, Point D)
 {
 	// Line AB represented as a1x + b1y = c1
-	double a1 = B.y - A.y;
-	double b1 = A.x - B.x;
-	double c1 = a1 * (A.x) + b1 * (A.y);
+	float a1 = B.y - A.y;
+	float b1 = A.x - B.x;
+	float c1 = a1 * (A.x) + b1 * (A.y);
 
 	// Line CD represented as a2x + b2y = c2
-	double a2 = D.y - C.y;
-	double b2 = C.x - D.x;
-	double c2 = a2 * (C.x) + b2 * (C.y);
+	float a2 = D.y - C.y;
+	float b2 = C.x - D.x;
+	float c2 = a2 * (C.x) + b2 * (C.y);
 
-	double determinant = a1 * b2 - a2 * b1; // determinant for a 2 dimensional matrix by Cramer's rule is calculated here
+	float determinant = a1 * b2 - a2 * b1; // determinant for a 2 dimensional matrix by Cramer's rule is calculated here
 
 	if (determinant == 0)	//If deteminant is 0 the lines are parallel and hence the point of intersection is set as inf, inf
 	{
-		Point p1(DBL_MAX, DBL_MAX);
+		Point p1(FLT_MAX, FLT_MAX);
 		return p1;
 	}
 	else					// if the lines aren't parallel, the x and y cordinates are calculated by the following equation
 	{
-		double x = (b2 * c1 - b1 * c2) / determinant;
-		double y = (a1 * c2 - a2 * c1) / determinant;
+		float x = (b2 * c1 - b1 * c2) / determinant;
+		float y = (a1 * c2 - a2 * c1) / determinant;
 		Point p1(x, y);
 		return p1;
 	}
